pem_platform/pyproject.toml
[build-system]
requires = ["hatchling>=1.24.2"]
build-backend = "hatchling.build"

[project]
name = "pem-platform"
version = "0.1.0"
description = "Process Engineering & DoE + Surrogate Modeling (GPR) + Retrieval (RAG-style) toolkit"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "Your Team", email = "team@example.com" }
]
keywords = ["DoE", "process-engineering", "gaussian-process", "bayesian-optimization", "RAG"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "scikit-learn>=1.3",
  "statsmodels>=0.14",
  "pyDOE2>=1.3.0",
  "matplotlib>=3.8",
  "pydantic>=2.6",
  "PyYAML>=6.0",
  "typer>=0.12",
  "rich>=13.7",
  "joblib>=1.3",
]

[project.optional-dependencies]
# Optional: higher-quality embeddings for retrieval.
embeddings = [
  "sentence-transformers>=3.0.0",
]

# Optional: Bayesian optimization via Ax (BoTorch backend)
ax = [
  "ax-platform>=0.3.6",
]

# Dev tooling
dev = [
  "pytest>=8.0",
  "ruff>=0.6",
  "build>=1.2",
]

[project.scripts]
pem = "pem.cli:app"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
addopts = "-q"

[tool.hatch.build.targets.wheel]
packages = ["src/pem"]
